Create sketch_ftp
FTP code

If FTPWRITE is defined, the sketch uploads the file to the FTP server. If not, it downloads the file to the SD card.

Change the network settings and the user/password for the server.

Press 'f' to upload/download the file. Press 'r' to read the current contents of the file on the SD

WARNING! Each device on a network must have a unique mac address. If you are using more than one ethernet shield on a network, you must insure all mac addresses are unique. No duplicates!